requirements.txt
python-telegram-bot
numpy
matplotlib
